nothing

+++ app/Http/Controllers/Admin/UserController.php
@@ -1,11 +1,13 @@
 <?php
 
 namespace App\Http\Controllers\Admin;
+
 use App\User;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Auth;
-use App\Http\Helpers\ControllerHelper;
+use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\MessageBag;
 
 class UserController extends Controller
 {
@@ -14,8 +16,10 @@ class UserController extends Controller
     {
         if (Auth::check()) {
             $users = User::orderBy('id', 'asc')->get();
-            return view('admin/user/index',
-                ['users' => $users,]);
+            return view('admin/user/index', [
+                'users' => $users,
+                'isAdmin' => Auth::user()->mode == 0,
+            ]);
         }
         return redirect('admin/login');
     }
@@ -24,8 +28,10 @@ class UserController extends Controller
     public function create()
     {
         if (Auth::check()) {
-
-            return view('admin/user/create');
+            if (Auth::user()->mode == 0) {
+                return view('admin/user/create');
+            }
+            return redirect('admin/users');
         }
 
         return redirect('admin/login');
@@ -35,13 +41,28 @@ class UserController extends Controller
     public function store(Request $request)
     {
         if (Auth::check()) {
-            $input = $request->all();          
-            $user = new User();
-            $user->name = $input['name'];
-            $user->email = $input['email'];
-            $user->password = $input['password'];
-            $user->mode = $input['mode'];
-            $user->save();
+            $data = $request->all();
+
+            $this->validator($data);
+
+            $user = User::where('email', $data['email'])->first();
+
+            if ($user) {
+                $errors = new MessageBag();
+
+                // add your error messages:
+                $errors->add('error', 'Email ' . $data['email'] . ' đã được đăng ký!');
+
+                return redirect('admin/users/create')->withErrors($errors);
+            }
+
+            User::create([
+                'name' => $data['name'],
+                'email' => $data['email'],
+                'password' => bcrypt($data['password']),
+                'mode' => $data['mode']
+            ]);
+
             return redirect('/admin/users');
         }
 
@@ -64,8 +85,11 @@ class UserController extends Controller
     public function edit($id)
     {
         if (Auth::check()) {
-            $user = User::find($id);
-            return view('admin/user/edit',['user'=>$user,]);
+            if (Auth::user()->mode == 0) {
+                $user = User::find($id);
+                return view('admin/user/edit', ['user' => $user,]);
+            }
+            return redirect('admin/users');
         }
 
         return redirect('admin/login');
@@ -75,9 +99,19 @@ class UserController extends Controller
     public function update(Request $request, $id)
     {
         if (Auth::check()) {
-            $input = $request->all();
-            $user = User::find($id);
-            $user->update($input);
+            if (Auth::user()->mode == 0) {
+                $input = $request->all();
+
+                $user = User::find($id);
+
+                $user->name = $input['name'];
+                if (!is_null($input['password'])) {
+                    $user->password = $input['password'];
+                }
+                $user->email = $input['email'];
+                $user->mode = $input['mode'];
+                $user->save;
+            }
             return redirect('admin/users');
         }
 
@@ -85,16 +119,30 @@ class UserController extends Controller
     }
 
     // DELETE /users/{user}
-    public function destroy()
+    public function destroy($id)
     {
         if (Auth::check()) {
-            User::where('id',$id)->delete();
-            $users = User::orderBy('id', 'asc')->get();
-            return view('admin/user/index', [
-                'users' => $users,
-            ]); 
+            if (Auth::user()->mode == 0 && Auth::user()->id != $id) {
+                User::where('id', $id)->delete();
+            }
+            return response()->json([
+                'status' => 'OK',
+            ]);
         }
 
-        return redirect('admin/login');
+        return response()->json([
+            'status' => 'ERROR',
+            'message' => 'Ban phai dang nhap',
+        ]);
+    }
+
+    protected function validator(array $data)
+    {
+        return Validator::make($data, [
+            'name' => 'required|string|max:255',
+            'email' => 'required|string|email|max:255|unique:users',
+            'password' => 'required|string|min:6|confirmed',
+            'mode' => 'required',
+        ]);
     }
 }
